refactor: move exceptions to infra/exceptions for better cohesion
